api/requirements.txt
pydantic-ai-slim[ag-ui]==0.0.19
httpx>=0.27
pydantic>=2.10
openai>=1.56
typing-extensions>=4.12